Implemented true momentum mechanic
Momentum is now its own score different from the overall score determined from accuracy (which still acts as the final score). It benefits from correct answers and opponent missed positives and wrong answers, and also has a bonus score based on round number if the user submits a perfect answer. If the momentum bar completely favors one side (given both sides have a chance to score points) it will end the game in favor of that side, and it gets easier as the game goes on.